requirements.txt
fastapi
uvicorn
python-docx
ollama
jinja2
python-multipart